Drop excess type annotations, conversions

+++ src/AIAPC2020/env.jl
@@ -123,7 +123,7 @@ end
 
 Return the reward for the current state for player `p` as an integer. If the episode is done, return the profit, else return `0`.
 """
-function RLBase.reward(env::AIAPCEnv, p::Int)::Float64
+function RLBase.reward(env::AIAPCEnv, p::Int)
     profit_array = env.profit_array
     memory_index = env.memory[1]
     return _reward(
@@ -135,7 +135,7 @@ end
 
 function _reward(profit::Array{Float64,3},
     memory_index::CartesianIndex{2},
-    p::Int)::Float64
+    p::Int)
 
     return profit[memory_index, p]
 end
